Refactor: Remove comparePassword.js and move methods to methodHelper.js to provide a modular file structure and dry principles
