--- src/helpers/comparePassword.js
@@ -1,13 +1,0 @@
-"use strict";
-
-/* ------------------------------------------------- */
-/*                  SOULJOURNEY API                  */
-/* ------------------------------------------------- */
-
-const bcrypt = require("bcryptjs");
-
-const comparePassword = async (candidatePassword, userPassword) => {
-  return await bcrypt.compare(candidatePassword, userPassword);
-};
-
-module.exports = comparePassword;
