+++ src/models/user.js
@@ -3,14 +3,17 @@
                     SOULJOURNEY API  
 ------------------------------------------------------- */
 
-const crypto = require("crypto");
 const { mongoose } = require("../configs/dbConnection");
 const validator = require("validator");
 const validatePassword = require("../helpers/validatePassword");
 const bcrypt = require("bcryptjs");
-const resetTokenHash = require("../helpers/resetTokenHash");
 const uniqueValidator = require("mongoose-unique-validator");
-const comparePassword = require("../helpers/comparePassword");
+const {
+  markAsVerified,
+  comparePassword,
+  createPasswordResetToken,
+  createVerificationCode,
+} = require("../helpers/methdodHelper");
 
 /* ------------------------------------------------------- *
 User Model requirements
@@ -129,43 +132,9 @@ UserSchema.pre("save", async function (next) {
   next();
 });
 
-// Method to mark the user's email as verified
-UserSchema.methods.markAsVerified = async function () {
-  this.isEmailVerified = true;
-  await this.save({ validateBeforeSave: false });
-};
-
-// Method to check if the provided password matches the hashed password in the database
+UserSchema.methods.markAsVerified = markAsVerified;
 UserSchema.methods.correctPassword = comparePassword;
-
-/*
- * This method generates a password reset token for a user.
- * It creates a random token, hashes it for security, and sets an expiration time.
- * The raw token is returned to be sent to the user via email.
- */
-UserSchema.methods.createPasswordResetToken = function () {
-  // Generates a random 32-byte reset token and converts it to a hexadecimal string
-  const resetToken = crypto.randomBytes(32).toString("hex");
-
-  // Hashes the reset token and stores it in the database for security
-  this.passwordResetToken = resetTokenHash(resetToken);
-
-  // Sets an expiration time for the reset token (10 minutes from now)
-  this.passwordResetExpires = Date.now() + 10 * 60 * 1000;
-
-  // Returns the raw reset token (not hashed) to be sent to the user
-  return resetToken;
-};
-
-// Method to create a verification code and set an expiration time
-UserSchema.methods.createVerificationCode = function () {
-  const verificationCode = Math.floor(100000 + Math.random() * 900000);
-
-  this.verificationCode = verificationCode;
-
-  this.verificationCodeExpires = Date.now() + 10 * 60 * 1000;
-
-  return verificationCode;
-};
+UserSchema.methods.createPasswordResetToken = createPasswordResetToken;
+UserSchema.methods.createVerificationCode = createVerificationCode;
 
 module.exports = mongoose.model("User", UserSchema);
